Restructure the project with a folder name 'pages' with the components home and offers which are pages.
