+++ src/app/app.module.ts
@@ -3,12 +3,12 @@ import { BrowserModule } from '@angular/platform-browser';
 
 import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
-import { HomeComponent } from './components/home/home.component';
+import { HomeComponent } from './pages/home/home.component';
 import { HeaderComponent } from './components/header/header.component';
 import { FooterComponent } from './components/footer/footer.component';
 import { EventComponent } from './components/event/event.component';
 import { OfferComponent } from './components/offer/offer.component';
-import { OffersPageComponent} from './components/offers-page/offers-page.component';
+import { OffersPageComponent} from './pages/offers/offers.component';
 import { HttpClientModule } from '@angular/common/http';
 import { ShoppingCartComponent } from "./components/shopping-cart/shopping-cart.component";
