+++ src/app/pages/offers/offers.component.html
